+++ .github/actions/generate-conventional-release-notes/closest-version.js
@@ -1,29 +1,17 @@
 const semver = require('semver');
-const fs = require('fs');
-const gitSemverTags = require('git-semver-tags');
+const semverTags = require('../helpers/git-semver-tags');
 
-const packageJson = JSON.parse(fs.readFileSync('./package.json', 'utf8'));
+const currentVersion = require('../helpers/current-version');
 
-const semverTags = (maxVersion) => {
-    return new Promise((resolve, reject) => {
-        gitSemverTags({ tagPrefix: 'v', skipUnstable: false }, function (err, result) {
-            if (err) {
-                reject(err);
-            } else {
-                const firstLessIndex = result.findIndex((v) => semver.lt(v, maxVersion));
-                resolve(result.slice(firstLessIndex, result.length));
-            }
-        });
-    });
-};
+const previousVersionPredicateForPrerelease = (v) => semver.gt(currentVersion, v) && v !== `v${currentVersion}`;
+const previousVersionPredicateForRelease = (v) => !semver.prerelease(v) && previousVersionPredicateForPrerelease(v);
 
 module.exports = async () => {
-    const isPrerelease = !!semver.prerelease(packageJson.version);
-    const availableVersions = await semverTags(packageJson.version);
-    let closestIndex = 0;
-    if (!isPrerelease) {
-        closestIndex = availableVersions.findIndex((v) => !semver.prerelease(v));
-    }
+    const isPrerelease = !!semver.prerelease(currentVersion);
+    const availableVersions = await semverTags(currentVersion);
+    const closestIndex = availableVersions.findIndex(
+        isPrerelease ? previousVersionPredicateForPrerelease : previousVersionPredicateForRelease
+    );
     return {
         closest: availableVersions[closestIndex],
         tagsTillClosest: availableVersions.slice(0, closestIndex)