+++ CHANGELOG.md
@@ -1,3 +1,72 @@
+# Change Log
+
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.43.0-rc.10](https://github.com/SAP/fundamental-ngx/compare/v0.43.0-rc.9...v0.43.0-rc.10) (2023-07-13)
+
+
+### Bug Fixes
+
+* **platform:** table console error ([#10163](https://github.com/SAP/fundamental-ngx/issues/10163)) ([3228343](https://github.com/SAP/fundamental-ngx/commit/3228343906f527c1f2b0455ba582af6bc6f7d91f))
+
+
+
+
+
+# [0.43.0-rc.9](https://github.com/SAP/fundamental-ngx/compare/v0.43.0-rc.8...v0.43.0-rc.9) (2023-07-13)
+
+
+### Bug Fixes
+
+* **docs,core,platform:** various fixes from deffect hunt ([#10168](https://github.com/SAP/fundamental-ngx/issues/10168)) ([00b858b](https://github.com/SAP/fundamental-ngx/commit/00b858bcfb1a1e6963c36a7d768fe9dcf100a135))
+
+
+
+
+
+# [0.43.0-rc.8](https://github.com/SAP/fundamental-ngx/compare/v0.43.0-rc.7...v0.43.0-rc.8) (2023-07-13)
+
+
+### Bug Fixes
+
+* **platform:** fixed table sort rules update ([#10167](https://github.com/SAP/fundamental-ngx/issues/10167)) ([457bfac](https://github.com/SAP/fundamental-ngx/commit/457bfac0dab855012614a02c802366c296479ab5))
+
+
+
+
+
+# [0.43.0-rc.7](https://github.com/SAP/fundamental-ngx/compare/v0.43.0-rc.6...v0.43.0-rc.7) (2023-07-12)
+
+
+### Bug Fixes
+
+* **core:** carousel not updating when slide count changes ([#10153](https://github.com/SAP/fundamental-ngx/issues/10153)) ([32e0eae](https://github.com/SAP/fundamental-ngx/commit/32e0eaeb96de4e31f82cbd23f11b0dc79eccc3c0))
+
+
+
+
+
+# [0.43.0-rc.6](https://github.com/SAP/fundamental-ngx/compare/v0.43.0-rc.5...v0.43.0-rc.6) (2023-07-12)
+
+**Note:** Version bump only for package fundamental-ngx
+
+
+
+
+
+# [0.43.0-rc.5](https://github.com/SAP/fundamental-ngx/compare/v0.43.0-rc.4...v0.43.0-rc.5) (2023-07-12)
+
+**Note:** Version bump only for package fundamental-ngx
+
+
+
+
+
+# [0.43.0-rc.4](https://github.com/SAP/fundamental-ngx/compare/v0.43.0-rc.3...v0.43.0-rc.4) (2023-07-11)
+
+
+
 # [0.43.0-rc.3](https://github.com/SAP/fundamental-ngx/compare/v0.43.0-rc.2...v0.43.0-rc.3) (2023-07-10)